SQLite only supports a few `ALTER TABLE` commands

+++ src/Statements/Alter/DropIndex.php
@@ -24,15 +24,20 @@
          */
         public function toSql(IDriver $driver): string
         {
-			if ($this->index === NULL) { // PRIMARY KEY
-				if ($driver instanceof Drivers\MysqlDriver) {
-					return 'DROP PRIMARY KEY';
+            if ($driver->modifyColumn ?? true) {
+                if ($this->index === NULL) { // PRIMARY KEY
+                    if ($driver instanceof Drivers\MysqlDriver) {
+                        return 'DROP PRIMARY KEY';
 
-				} else {
-					throw new NotImplementedException('Drop of primary key is not implemented for driver ' . get_class($driver) . '.');
-				}
-			}
+                    } else {
+                        throw new NotImplementedException('Drop of primary key is not implemented for driver ' . get_class($driver) . '.');
+                    }
+                }
 
-			return 'DROP INDEX ' . $driver->escapeIdentifier($this->index);
+                return 'DROP INDEX ' . $driver->escapeIdentifier($this->index);
+            }
+            else {
+                throw new NotImplementedException('Drop key is not implemented for driver ' . get_class($driver) . '.');
+            }
 		}
 	}
